+++ apps/cowswap-frontend/src/common/types/bridge.ts
@@ -2,7 +2,6 @@ import { TokenWithLogo } from '@cowprotocol/common-const'
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 export interface BridgeEstimatedAmounts {
-  sellAmount: CurrencyAmount<Currency>
   expectedToReceiveAmount: CurrencyAmount<Currency>
   feeAmount: CurrencyAmount<Currency>
   minToReceiveAmount: CurrencyAmount<Currency>