+++ apps/cowswap-frontend/src/common/updaters/orders/PendingOrdersUpdater.ts
@@ -50,6 +50,7 @@ import { getUiOrderType } from 'utils/orderUtils/getUiOrderType'
 
 import { fetchAndClassifyOrder } from './utils'
 
+import { BridgeOrdersStateSerialized } from '../../../entities/bridgeOrders/state/bridgeOrdersAtom'
 import { useBlockNumber } from '../../hooks/useBlockNumber'
 import { removeOrdersToCancelAtom } from '../../hooks/useMultipleOrdersCancellation/state'
 import { useTriggerTotalSurplusUpdateCallback } from '../../state/totalSurplusState'
@@ -166,6 +167,7 @@ interface UpdateOrdersParams {
   chainId: ChainId
   account: string
   isSafeWallet: boolean
+  bridgeOrdersMap: BridgeOrdersStateSerialized
   orders: Order[]
   // Actions
   addOrUpdateOrders: AddOrUpdateOrdersCallback
@@ -186,7 +188,9 @@ interface UpdateOrdersParams {
 interface HandlePresignedParams {
   presigned: EnrichedOrder[]
   orders: Order[]
+  bridgeOrdersMap: BridgeOrdersStateSerialized
   chainId: ChainId
+  account: string
   isSafeWallet: boolean
   presignOrders: PresignOrdersCallback
 }
@@ -194,7 +198,9 @@ interface HandlePresignedParams {
 function handlePresignedOrders({
   presigned,
   orders,
+  bridgeOrdersMap,
   chainId,
+  account,
   isSafeWallet,
   presignOrders,
 }: HandlePresignedParams): void {
@@ -223,8 +229,12 @@ function handlePresignedOrders({
     isSafeWallet,
   })
 
+  const bridgeOrders = bridgeOrdersMap[chainId]?.[account.toLowerCase()]
+
   newlyPreSignedOrders.forEach((order) => {
-    emitPresignedOrderEvent({ chainId, order })
+    const bridgeOrder = bridgeOrders?.find((i) => i.orderUid === order.uid)
+
+    emitPresignedOrderEvent({ chainId, order, bridgeOrder })
   })
 }
 
@@ -234,6 +244,7 @@ async function _updateOrders({
   account,
   chainId,
   orders,
+  bridgeOrdersMap,
   isSafeWallet,
   // Actions
   addOrUpdateOrders,
@@ -281,7 +292,7 @@ async function _updateOrders({
     { fulfilled: [], expired: [], cancelled: [], unknown: [], presigned: [], pending: [], presignaturePending: [] },
   )
 
-  handlePresignedOrders({ presigned, orders, chainId, isSafeWallet, presignOrders })
+  handlePresignedOrders({ presigned, orders, bridgeOrdersMap, chainId, account, isSafeWallet, presignOrders })
 
   if (expired.length > 0) {
     expireOrdersBatch({
@@ -451,6 +462,10 @@ export function PendingOrdersUpdater(): null {
   const allTransactions = useAllTransactions()
   const getSafeTxInfo = useGetSafeTxInfo()
   const bridgeOrdersMap = useBridgeOrdersSerializedMap()
+  const bridgeOrdersMapRef = useRef(bridgeOrdersMap)
+  useEffect(() => {
+    bridgeOrdersMapRef.current = bridgeOrdersMap
+  }, [bridgeOrdersMap])
   const markPollComplete = useCallback(
     (targetChainId: ChainId) => {
       if (!chainId || targetChainId !== chainId) {
@@ -501,6 +516,7 @@ export function PendingOrdersUpdater(): null {
           account,
           chainId,
           isSafeWallet,
+          bridgeOrdersMap: bridgeOrdersMapRef.current,
           orders: pendingRef.current.filter((order) => getUiOrderType(order) === uiOrderType),
           addOrUpdateOrders,
           fulfillOrdersBatch,