+++ apps/cowswap-frontend/src/modules/bridge/hooks/useQuoteBridgeContext.ts
@@ -1,16 +1,13 @@
 import { useMemo } from 'react'
 
-import { useWalletInfo } from '@cowprotocol/wallet'
-
 import { useBridgeSupportedNetwork } from 'entities/bridgeProvider'
 
-import { useDerivedTradeState } from 'modules/trade'
 import { useEstimatedBridgeBuyAmount } from 'modules/trade'
 import { useTradeQuote } from 'modules/tradeQuote'
-import { BRIDGE_QUOTE_ACCOUNT } from 'modules/tradeQuote'
 import { useUsdAmount } from 'modules/usdAmount'
 
 import { useBridgeQuoteAmounts } from './useBridgeQuoteAmounts'
+import { useBridgeQuoteRecipient } from './useBridgeQuoteRecipient'
 
 import { QuoteBridgeContext } from '../types'
 
@@ -18,29 +15,28 @@ export function useQuoteBridgeContext(): QuoteBridgeContext | null {
   const { bridgeQuote } = useTradeQuote()
 
   const quoteAmounts = useBridgeQuoteAmounts()
-
-  const { expectedToReceiveAmount: buyAmount } = useEstimatedBridgeBuyAmount() || {}
   const bridgeFee = quoteAmounts?.bridgeFee
+  const estimatedAmounts = useEstimatedBridgeBuyAmount()
+  const buyAmount = estimatedAmounts?.expectedToReceiveAmount
+  const expectedToReceiveAmount = useMemo(() => {
+    if (!buyAmount || !bridgeFee) return null
 
-  const expectedToReceive = useMemo(() => {
-    return buyAmount && bridgeFee ? buyAmount.subtract(bridgeFee) : null
+    return buyAmount.subtract(bridgeFee)
   }, [buyAmount, bridgeFee])
 
   const { value: buyAmountUsd } = useUsdAmount(buyAmount)
-  const { value: bridgeMinDepositAmountUsd } = useUsdAmount(quoteAmounts?.swapMinReceiveAmount)
-  const { value: expectedToReceiveUsd } = useUsdAmount(expectedToReceive)
+  const { value: bridgeMinDepositAmountUsd } = useUsdAmount(quoteAmounts?.bridgeMinReceiveAmount)
+  const { value: expectedToReceiveUsd } = useUsdAmount(expectedToReceiveAmount)
+  const { value: bridgeMinReceiveAmountUsd } = useUsdAmount(quoteAmounts?.bridgeMinReceiveAmount)
 
   const targetChainId = quoteAmounts?.bridgeMinReceiveAmount.currency.chainId
   const destChainData = useBridgeSupportedNetwork(targetChainId)
 
-  const { account } = useWalletInfo()
-  const tradeState = useDerivedTradeState()
-
   const expectedFillTimeSeconds = bridgeQuote?.expectedFillTimeSeconds
-  const recipient = tradeState?.recipient || account || BRIDGE_QUOTE_ACCOUNT
+  const recipient = useBridgeQuoteRecipient()
 
   return useMemo(() => {
-    if (!quoteAmounts || !recipient || !buyAmount) return null
+    if (!quoteAmounts?.swapExpectedReceive || !recipient || !buyAmount) return null
 
     if (!destChainData) return null
 
@@ -49,14 +45,15 @@ export function useQuoteBridgeContext(): QuoteBridgeContext | null {
       bridgeFee: quoteAmounts.bridgeFee,
       estimatedTime: expectedFillTimeSeconds || null,
       recipient,
-      sellAmount: quoteAmounts.swapBuyAmount,
+      sellAmount: quoteAmounts.swapExpectedReceive,
       buyAmount: buyAmount,
       bridgeReceiverOverride: bridgeQuote?.bridgeReceiverOverride || null,
       bridgeMinDepositAmount: quoteAmounts.swapMinReceiveAmount,
       bridgeMinReceiveAmount: quoteAmounts.bridgeMinReceiveAmount,
+      bridgeMinReceiveAmountUsd,
       bridgeMinDepositAmountUsd,
       buyAmountUsd,
-      expectedToReceive,
+      expectedToReceive: expectedToReceiveAmount,
       expectedToReceiveUsd,
     }
   }, [
@@ -68,7 +65,8 @@ export function useQuoteBridgeContext(): QuoteBridgeContext | null {
     buyAmountUsd,
     bridgeMinDepositAmountUsd,
     bridgeQuote?.bridgeReceiverOverride,
-    expectedToReceive,
+    expectedToReceiveAmount,
     expectedToReceiveUsd,
+    bridgeMinReceiveAmountUsd,
   ])
 }